test: extract layoutAsserts.ts; drop .test.ts-as-helper anti-pattern
layoutPacker.test.ts imported the four assert* helpers from
layout.test.ts, which caused vitest to register layout.test.ts's
describe blocks twice (once per importer worker context), inflating
the suite total by 74 ghost duplicates (1935 → 1861 true unique tests).

Move assertNoOverlap, assertStemOrder, assertTreeRespecting, and
assertTJunctionsValid (plus their private rect/join helpers) to
app/tests/_helpers/layoutAsserts.ts. Both layout.test.ts and
layoutPacker.test.ts now import from the helper module.
